snapshot of defaults for 2020-11-15 results
in which i made the data linearly separable and threw away different % of positives to skew the ratio of positives to negatives

result: the more positives kept, the faster the rate of convergence for b_error. my guess is that this is due to decreasing the overall size of the dataset for the low-positive-ratio runs. will try next to keep the dataset size fixed by adding in negatives or normalizing or something.