bugfix: compute b error
The previous expression for b error didn't make sense because it was not scale invariant. For example, if b = [1, 1] and bhat = [2, 2], b_error given by the l2 norm of the difference scales with the difference, even though the vectors have the same direction. Rather than dividing the l2 norm difference by the l2 norm of b (which accomplishes nothing), both vectors should be unit-normalized before computing the difference vector and return its l2 norm.

Test plan: pytest plus some sanity checks. Now (finally) the b_error looks sane, in that for linearly separable data with vanilla SGD, it is not increasing while the validation and/or train loss is decreasing.